Fix GCC compiler warning (possible loss of precision).
git-svn-id: file:///.data/svn_convert/svn@7081 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/yahoo/yahooclient.cpp
@@ -879,7 +879,7 @@ void YahooClient::processStatus(unsigned short service, const char *id,
     YahooUserData *data = findContact(id, NULL, contact);
     if (data == NULL)
         return;
-    unsigned state = 0;
+    unsigned long state = 0;
     unsigned away  = 0;
     unsigned idle  = 0;
     if (_state)
